@@ -2,12 +2,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import '../../../app/theme.dart';
+import '../../../core/utils/responsive.dart';
+import '../../widgets/reset_button.dart';
 
 class SaveLoadScreen extends ConsumerWidget {
   const SaveLoadScreen({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    Responsive.init(context);
     // 샘플 세이브 데이터
     final saves = [
       {
@@ -38,64 +41,73 @@ class SaveLoadScreen extends ConsumerWidget {
           onPressed: () => context.go('/'),
         ),
       ),
-      body: ListView.builder(
-        padding: const EdgeInsets.all(16),
-        itemCount: saves.length,
-        itemBuilder: (context, index) {
-          final save = saves[index];
-          final isEmpty = save['team'] == null;
+      body: Stack(
+        children: [
+          ListView.builder(
+            padding: const EdgeInsets.all(16),
+            itemCount: saves.length,
+            itemBuilder: (context, index) {
+              final save = saves[index];
+              final isEmpty = save['team'] == null;
 
-          return _SaveSlotCard(
-            slot: save['slot'] as int,
-            team: save['team'] as String?,
-            season: save['season'] as int?,
-            rank: save['rank'] as int?,
-            date: save['date'] as String?,
-            isEmpty: isEmpty,
-            onLoad: isEmpty
-                ? null
-                : () {
-                    // TODO: 로드 처리
-                    context.go('/main');
-                  },
-            onSave: () {
-              // TODO: 세이브 처리
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text('슬롯 ${save['slot']}에 저장되었습니다')),
+              return _SaveSlotCard(
+                slot: save['slot'] as int,
+                team: save['team'] as String?,
+                season: save['season'] as int?,
+                rank: save['rank'] as int?,
+                date: save['date'] as String?,
+                isEmpty: isEmpty,
+                onLoad: isEmpty
+                    ? null
+                    : () {
+                        // TODO: 로드 처리
+                        context.go('/main');
+                      },
+                onSave: () {
+                  // TODO: 세이브 처리
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    SnackBar(content: Text('슬롯 ${save['slot']}에 저장되었습니다')),
+                  );
+                },
+                onDelete: isEmpty
+                    ? null
+                    : () {
+                        // TODO: 삭제 처리
+                        showDialog(
+                          context: context,
+                          builder: (context) => AlertDialog(
+                            title: const Text('삭제 확인'),
+                            content: Text('슬롯 ${save['slot']}의 데이터를 삭제하시겠습니까?'),
+                            actions: [
+                              TextButton(
+                                onPressed: () => Navigator.pop(context),
+                                child: const Text('취소'),
+                              ),
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.pop(context);
+                                  ScaffoldMessenger.of(context).showSnackBar(
+                                    const SnackBar(content: Text('삭제되었습니다')),
+                                  );
+                                },
+                                child: const Text(
+                                  '삭제',
+                                  style: TextStyle(color: Colors.red),
+                                ),
+                              ),
+                            ],
+                          ),
+                        );
+                      },
               );
             },
-            onDelete: isEmpty
-                ? null
-                : () {
-                    // TODO: 삭제 처리
-                    showDialog(
-                      context: context,
-                      builder: (context) => AlertDialog(
-                        title: const Text('삭제 확인'),
-                        content: Text('슬롯 ${save['slot']}의 데이터를 삭제하시겠습니까?'),
-                        actions: [
-                          TextButton(
-                            onPressed: () => Navigator.pop(context),
-                            child: const Text('취소'),
-                          ),
-                          TextButton(
-                            onPressed: () {
-                              Navigator.pop(context);
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                const SnackBar(content: Text('삭제되었습니다')),
-                              );
-                            },
-                            child: const Text(
-                              '삭제',
-                              style: TextStyle(color: Colors.red),
-                            ),
-                          ),
-                        ],
-                      ),
-                    );
-                  },
-          );
-        },
+          ),
+          Positioned(
+            bottom: 80.sp,
+            left: 16.sp,
+            child: const ResetButton(),
+          ),
+        ],
       ),
     );
   }
