@@ -3,6 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import '../../../app/theme.dart';
 import '../../../core/constants/team_data.dart';
+import '../../widgets/reset_button.dart';
 
 class TeamSelectScreen extends ConsumerStatefulWidget {
   const TeamSelectScreen({super.key});
@@ -24,106 +25,117 @@ class _TeamSelectScreenState extends ConsumerState<TeamSelectScreen> {
           onPressed: () => context.go('/'),
         ),
       ),
-      body: Column(
+      body: Stack(
         children: [
-          const Padding(
-            padding: EdgeInsets.all(16.0),
-            child: Text(
-              '운영할 게임단을 선택하세요',
-              style: TextStyle(
-                fontSize: 18,
-                color: AppTheme.textSecondary,
-              ),
-            ),
-          ),
-
-          // 팀 그리드
-          Expanded(
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16.0),
-              child: GridView.builder(
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 3,
-                  childAspectRatio: 0.85,
-                  crossAxisSpacing: 12,
-                  mainAxisSpacing: 12,
-                ),
-                itemCount: TeamData.teams.length,
-                itemBuilder: (context, index) {
-                  final team = TeamData.teams[index];
-                  final isSelected = selectedTeamId == team['id'];
-
-                  return _TeamCard(
-                    team: team,
-                    isSelected: isSelected,
-                    onTap: () {
-                      setState(() {
-                        selectedTeamId = team['id'];
-                      });
-                    },
-                  );
-                },
-              ),
-            ),
-          ),
-
-          // 선택된 팀 정보
-          if (selectedTeamId != null) ...[
-            Container(
-              width: double.infinity,
-              padding: const EdgeInsets.all(16),
-              color: AppTheme.cardBackground,
-              child: Column(
-                children: [
-                  Text(
-                    TeamData.teams.firstWhere(
-                      (t) => t['id'] == selectedTeamId,
-                    )['name'],
-                    style: const TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-                  Text(
-                    '에이스: ${TeamData.teams.firstWhere((t) => t['id'] == selectedTeamId)['ace']}',
-                    style: const TextStyle(
-                      color: AppTheme.textSecondary,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-
-          // 확정 버튼
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: SizedBox(
-              width: double.infinity,
-              height: 56,
-              child: ElevatedButton(
-                onPressed: selectedTeamId == null
-                    ? null
-                    : () {
-                        // TODO: 선택한 팀 저장
-                        context.go('/main');
-                      },
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: AppTheme.accentGreen,
-                  foregroundColor: Colors.black,
-                  disabledBackgroundColor: AppTheme.cardBackground,
-                ),
-                child: const Text(
-                  '게임 시작',
+          Column(
+            children: [
+              const Padding(
+                padding: EdgeInsets.all(16.0),
+                child: Text(
+                  '운영할 게임단을 선택하세요',
                   style: TextStyle(
                     fontSize: 18,
-                    fontWeight: FontWeight.bold,
+                    color: AppTheme.textSecondary,
                   ),
                 ),
               ),
-            ),
+
+              // 팀 그리드
+              Expanded(
+                child: Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  child: GridView.builder(
+                    gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                      crossAxisCount: 3,
+                      childAspectRatio: 0.85,
+                      crossAxisSpacing: 12,
+                      mainAxisSpacing: 12,
+                    ),
+                    itemCount: TeamData.teams.length,
+                    itemBuilder: (context, index) {
+                      final team = TeamData.teams[index];
+                      final teamId = team['id'] as String;
+                      final isSelected = selectedTeamId == teamId;
+
+                      return _TeamCard(
+                        team: team,
+                        isSelected: isSelected,
+                        onTap: () {
+                          setState(() {
+                            selectedTeamId = teamId;
+                          });
+                        },
+                      );
+                    },
+                  ),
+                ),
+              ),
+
+              // 선택된 팀 정보
+              if (selectedTeamId != null) ...[
+                Builder(
+                  builder: (context) {
+                    final selectedTeam = TeamData.teams.firstWhere(
+                      (t) => t['id'] == selectedTeamId,
+                    );
+                    return Container(
+                      width: double.infinity,
+                      padding: const EdgeInsets.all(16),
+                      color: AppTheme.cardBackground,
+                      child: Column(
+                        children: [
+                          Text(
+                            selectedTeam['name'] as String,
+                            style: const TextStyle(
+                              fontSize: 20,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                          const SizedBox(height: 4),
+                          Text(
+                            '에이스: ${selectedTeam['ace'] as String}',
+                            style: const TextStyle(
+                              color: AppTheme.textSecondary,
+                            ),
+                          ),
+                        ],
+                      ),
+                    );
+                  },
+                ),
+              ],
+
+              // 확정 버튼
+              Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: SizedBox(
+                  width: double.infinity,
+                  height: 56,
+                  child: ElevatedButton(
+                    onPressed: selectedTeamId == null
+                        ? null
+                        : () {
+                            // TODO: 선택한 팀으로 게임 시작
+                            context.go('/director-name');
+                          },
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: AppTheme.accentGreen,
+                      foregroundColor: Colors.black,
+                      disabledBackgroundColor: AppTheme.cardBackground,
+                    ),
+                    child: const Text(
+                      '게임 시작',
+                      style: TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ],
           ),
+          ResetButton.positioned(),
         ],
       ),
     );
@@ -143,6 +155,10 @@ class _TeamCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final teamColor = Color(team['color'] as int);
+    final teamName = team['name'] as String;
+    final teamShortName = team['shortName'] as String;
+
     return GestureDetector(
       onTap: onTap,
       child: AnimatedContainer(
@@ -151,15 +167,13 @@ class _TeamCard extends StatelessWidget {
           color: AppTheme.cardBackground,
           borderRadius: BorderRadius.circular(12),
           border: Border.all(
-            color: isSelected
-                ? Color(team['color'])
-                : Colors.transparent,
+            color: isSelected ? teamColor : Colors.transparent,
             width: 3,
           ),
           boxShadow: isSelected
               ? [
                   BoxShadow(
-                    color: Color(team['color']).withOpacity(0.3),
+                    color: teamColor.withOpacity(0.3),
                     blurRadius: 12,
                     spreadRadius: 2,
                   ),
@@ -174,18 +188,18 @@ class _TeamCard extends StatelessWidget {
               width: 60,
               height: 60,
               decoration: BoxDecoration(
-                color: Color(team['color']).withOpacity(0.2),
+                color: teamColor.withOpacity(0.2),
                 shape: BoxShape.circle,
                 border: Border.all(
-                  color: Color(team['color']),
+                  color: teamColor,
                   width: 2,
                 ),
               ),
               child: Center(
                 child: Text(
-                  team['shortName'],
+                  teamShortName,
                   style: TextStyle(
-                    color: Color(team['color']),
+                    color: teamColor,
                     fontWeight: FontWeight.bold,
                     fontSize: 14,
                   ),
@@ -195,7 +209,7 @@ class _TeamCard extends StatelessWidget {
             const SizedBox(height: 8),
             // 팀명
             Text(
-              team['name'],
+              teamName,
               textAlign: TextAlign.center,
               style: const TextStyle(
                 fontSize: 12,
